管理员可删除用户

+++ app/Http/Controllers/UsersController.php
@@ -82,7 +82,7 @@ class UsersController extends Controller
 
     public function destroy(User $user)
     {
-        $user->authorize('destroy', $user);
+        $this->authorize('destroy', $user);
         $user->delete();
         session()->flash('success', '成功删除用户！');
         return back();
